py_dnd/pyproject.toml
[tool.poetry]
name = "py-dnd"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
aiosqlite = "^0.20.0"
alembic = "^1.13.3"
asyncpg = "0.30.0"
greenlet = "^3.1.1"
fastapi = "^0.115.2"
loguru = "^0.7.2"
pydantic-settings = "^2.5.2"
python-multipart = "^0.0.9"
sqlalchemy = "^2.0.35"
uvicorn = "^0.30.6"
pandas = "^2.2.3"
aiohttp = "3.11.10"
python-keycloak = "^4.6.2"
fastapi-keycloak = "^1.0.11"

[tool.poetry.group.dev.dependencies]
bandit = "^1.7.9"
coverage = "^7.6.1"
mkdocs = "^1.6.1"
mkdocs-material = "^9.5.35"
mkdocstrings = "^0.26.1"
pylint = "^3.3.0"
pytest = "^8.3.3"
pytest-cov = "^5.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.plugins."poetry.application.plugin"]
export = "poetry_plugin_export.plugins:ExportApplicationPlugin"
